Files in order. Ready to start final phase.
