+++ app/src/main/java/com/staminapp/ui/execute/ExecuteRoutineScreen.kt
@@ -1,6 +1,5 @@
-package com.staminapp
+package com.staminapp.ui.execute
 
-import androidx.annotation.FloatRange
 import androidx.compose.animation.core.*
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.Image
@@ -21,7 +20,6 @@ import androidx.compose.material.icons.filled.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.PaintingStyle.Companion.Stroke
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.text.style.TextAlign
@@ -29,7 +27,8 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.TextUnit
-import androidx.navigation.compose.rememberNavController
+import com.staminapp.R
+import com.staminapp.ui.main.RatingBar
 import com.staminapp.ui.theme.StaminappAppTheme
 import java.util.concurrent.TimeUnit
