+++ app/src/main/java/com/staminapp/ui/main/Composables.kt
@@ -1,0 +1,113 @@
+package com.staminapp.ui.main
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import com.staminapp.R
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Star
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Brush
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import com.staminapp.Destination
+import com.staminapp.ui.theme.Gray
+
+@Composable
+fun RatingBar(
+    modifier: Modifier = Modifier,
+    rating: Int = 0,
+    stars: Int = 5,
+    starsColor: Color = Color.Yellow,
+) {
+    Row(modifier = modifier) {
+        repeat(rating) {
+            Icon(Icons.Filled.Star, contentDescription = null, tint = starsColor)
+        }
+        repeat(stars - rating) {
+            Icon(Icons.Filled.Star, contentDescription = null, tint = Gray)
+        }
+    }
+}
+
+@Composable
+fun CustomChip(
+    selected: Boolean,
+    text: String,
+    modifier: Modifier = Modifier
+) {
+    Surface(
+        color = when {
+            selected -> MaterialTheme.colors.primaryVariant
+            else -> Gray
+        },
+        contentColor = when {
+            selected -> Color.White
+            else -> MaterialTheme.colors.primaryVariant
+        },
+        shape = CircleShape,
+        modifier = modifier,
+    ) {
+        Text(
+            text = text.uppercase(),
+            textAlign = TextAlign.Center,
+            style = MaterialTheme.typography.body2,
+            modifier = Modifier.padding(12.dp, 8.dp)
+        )
+    }
+}
+
+@Composable
+fun RoutineCard(navController: NavController, modifier: Modifier = Modifier) {
+    Card(
+        modifier = modifier.clickable{navController.navigate(Destination.Routine.route)},
+        shape = RoundedCornerShape(15.dp),
+        elevation = 5.dp
+    ) {
+        Box(modifier = Modifier
+            .fillMaxWidth()
+            .height(160.dp)) {
+            Image(
+                painter = painterResource(id = R.drawable.tincho2),
+                contentDescription = "Routine",
+                contentScale = ContentScale.Crop
+            )
+            Box(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .background(
+                        Brush.verticalGradient(
+                            colors = listOf(
+                                Color.Transparent,
+                                Color.Black
+                            ),
+                            startY = 300f
+                        )
+                    )
+            )
+            Box(
+                modifier = Modifier
+                    .fillMaxSize()
+                    .padding(12.dp),
+                contentAlignment = Alignment.BottomStart
+            ) {
+                Text(text = "Routine Name",
+                    color = MaterialTheme.colors.background,
+                    overflow = TextOverflow.Ellipsis,
+                    maxLines = 2
+                )
+            }
+        }
+    }
+}
